requirements.txt
openai
rich
mcp
python-dotenv
matplotlib
pandas
numpy
requests
textblob
docker
psutil
scikit-learn
cryptography
watchdog
PyYAML
PyPDF2
